packages/config/pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = "Configuration management"
authors = ["CKTT Team <team@cktt.ai>"]
packages = [{include = "config"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
